voice/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
audioop-lts==0.2.2
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
coloredlogs==15.0.1
daily-python==0.14.2
distro==1.9.0
fastapi==0.116.1
flatbuffers==25.9.23
frozenlist==1.8.0
future==1.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.26.0
google-api-python-client==2.185.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-cloud-texttospeech==2.21.1
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
humanfriendly==10.0
idna==3.11
jiter==0.11.1
llvmlite==0.44.0
loguru==0.7.3
Markdown==3.9
mpmath==1.3.0
multidict==6.7.0
numba==0.61.2
numpy==2.1.3
onnxruntime==1.20.1
openai==1.57.4
packaging==25.0
pillow==11.0.0
pipecat-ai==0.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.10.6
pydantic_core==2.27.2
pyloudnorm==0.1.1
pyparsing==3.2.5
python-deepcompare==1.0.1
python-dotenv==1.1.1
requests==2.32.5
resampy==0.4.3
rsa==4.9.1
scipy==1.16.2
sniffio==1.3.1
starlette==0.47.3
sympy==1.14.0
tenacity==9.0.0
tqdm==4.67.1
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
websockets==13.1
yarl==1.22.0
